% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_classifier.R
\name{io_classifier}
\alias{io_classifier}
\title{Sort Catch Numbers Into Species Categories}
\usage{
io_classifier(
  data,
  species = Comm.Catch.Spp.List,
  year = NA,
  recall = T,
  tsn = NULL
)
}
\arguments{
\item{data}{A data frame that includes catch numbers in dollars and TSN.}

\item{year}{A numeric variable that can be used to filter for a specific year. Defaults to NA, which returns all years.}

\item{tsn}{A dataset that can be used in place of the API call. Defaults to NULL}

\item{list}{A list of lists that includes categories at the top level and TSN numbers at the second level. Defaults to Comm.Catch.Spp.List.}

\item{recal}{A binary variable that determines whether the API should be queried or a default value used. Defaults to TRUE}
}
\description{
This function uses itis_reclassify(), developed by Emily Markowitz, to create species category classifications and then sorts catch numbers into those classifications.
}
