% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_cleaner.R
\name{io_cleaner}
\alias{io_cleaner}
\title{Clean the Output From the IO Model}
\usage{
io_cleaner(impact, format = "national", xlsx = F, fp = fips, maxyr = 2018)
}
\arguments{
\item{impact}{A data frame that was output by io_calculator().}

\item{format}{A string variable with several options for specifying output format.}

\item{xlsx}{A boolean variable that exports the output as tabs in an xlsx file if True.}

\item{fp}{A data frame containing state names and fips codes. Defaults to standard with EFL = 12 and WFL = 12.5.}

\item{maxyr}{The catch year - only used for FEUS.}
}
\description{
This function presents the output from io_calculator() in a variety of specifiable formats.
}
